% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnbtn_intersect_effects.R
\name{rnbtn_intersect_effects}
\alias{rnbtn_intersect_effects}
\title{rnbtn_intersect_effects combines results from the total and unique counts and produces intersected .csv files
with specified fdr cutoffs.It also produces intersection plots}
\usage{
rnbtn_intersect_effects(TC_files, UC_files, path, fdrcutoff = 0.05)
}
\arguments{
\item{TC_files}{: A list of Total counts files from rnbtn_fdr_effects along with full path.
Each file should have columns of fdr,log2_coefficient,locus_tag,mean,controlmean}

\item{UC_files}{: A list of Unique counts files from rnbtn_fdr_effects along with full path.
Each file should have fdr,log2_coefficient,locus_tag,mean,controlmean}

\item{path:}{A path to write output files.}

\item{fdrcutoff:}{A cutoff for significant locus_tags.Default is 0.05}
}
\description{
rnbtn_intersect_effects combines results from the total and unique counts and produces intersected .csv files
with specified fdr cutoffs.It also produces intersection plots
}
