% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnbtn_simulate_data.R
\name{rnbtn_simulate_data}
\alias{rnbtn_simulate_data}
\title{rnbtn_simulate_data simulates the transposon count data for Total counts
 and Unique Counts in a nested experimental structure of strain/condition/slevel with replicates.
It simulates 4000 locus_tags for each replicate.}
\usage{
rnbtn_simulate_data(n_strain, n_condition, n_slevel, n_rep)
}
\arguments{
\item{n_rep}{= Number of replicates for each stress level}

\item{n_strain=}{Number of strains or genetic backgrounds  to simulate}

\item{n_condition=}{Number of experimental/environmental conditions to simulate
for each strain}

\item{n_slevel=}{Number of stress levels to simulate for each condition}
}
\description{
rnbtn_simulate_data simulates the transposon count data for Total counts
 and Unique Counts in a nested experimental structure of strain/condition/slevel with replicates.
It simulates 4000 locus_tags for each replicate.
}
\examples{
rnbtn_simulate_data(n_strain=2,n_condition=2,n_slevel=2,n_rep=2)
}
